Refactor data sample
